requirements.txt
attrs==25.3.0
Automat==25.4.16
better-profanity==0.7.0
certifi==2025.1.31
charset-normalizer==3.4.1
constantly==23.10.4
filelock==3.17.0
fsspec==2025.2.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
pyasn1==0.6.1
pydub==0.25.1
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
sympy==1.13.1
tiktoken==0.9.0
tomli==2.2.1
torch==2.6.0
tqdm==4.67.1
Twisted==24.11.0
typing_extensions==4.12.2
urllib3==2.3.0
zope.interface==7.1.0
